requirements.txt
aio-pika==9.5.5
aiormq==6.8.1
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
beanie==1.30.0
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comm==0.2.2
debugpy==1.8.15
decorator==5.2.1
dnspython==2.7.0
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.116.1
google-auth==2.40.3
google-genai==1.25.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.8.1
lazy-model==0.2.0
matplotlib-inline==0.1.7
motor==3.3.2
multidict==6.6.3
nest-asyncio==1.6.0
packaging==25.0
pamqp==3.3.0
parso==0.8.4
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.3.2
psutil==7.0.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pymongo==4.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0

pyzmq==27.0.0
requests==2.32.4
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.47.1
tenacity==8.5.0
tornado==6.5.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
wcwidth==0.2.13
websockets==15.0.1
yarl==1.20.1
